Made jit json deserialization return errors instead of exploding
