+++ crates/dataflow-jit/src/codegen/json/mod.rs
@@ -1,0 +1,10 @@
+mod deserialize;
+mod serialize;
+mod tests;
+
+pub use deserialize::{call_deserialize_fn, DeserializeJsonFn, DeserializeResult, JsonDeserConfig};
+pub use serialize::{JsonSerConfig, SerializeJsonFn};
+
+// The index of a column within a row
+// TODO: Newtyping for column indices within the layout interfaces
+type ColumnIdx = usize;
