+++ crates/dataflow-jit/src/facade/mod.rs
@@ -1,6 +1,8 @@
+mod tests;
+
 use crate::{
     codegen::{
-        json::{DeserializeJsonFn, JsonDeserConfig},
+        json::{call_deserialize_fn, DeserializeJsonFn, JsonDeserConfig},
         CodegenConfig, NativeLayout, NativeLayoutCache, VTable,
     },
     dataflow::{CompiledDataflow, JitHandle, RowInput, RowOutput},
@@ -23,7 +25,8 @@ use dbsp::{
 use rust_decimal::Decimal;
 use serde_json::{Deserializer, Value};
 use std::{
-    collections::BTreeMap, io::Read, mem::transmute, ops::Not, path::Path, thread, time::Instant,
+    collections::BTreeMap, error, io::Read, mem::transmute, ops::Not, path::Path, thread,
+    time::Instant,
 };
 
 // TODO: A lot of this still needs fleshing out, mainly the little tweaks that
@@ -69,11 +72,23 @@ pub struct JsonSetHandle {
 }
 
 impl JsonSetHandle {
-    pub fn push(&self, key: &[u8], weight: i32) -> Result<(), serde_json::Error> {
+    pub fn new(
+        handle: CollectionHandle<Row, i32>,
+        deserialize_fn: DeserializeJsonFn,
+        vtable: &'static VTable,
+    ) -> Self {
+        Self {
+            handle,
+            deserialize_fn,
+            vtable,
+        }
+    }
+
+    pub fn push(&self, key: &[u8], weight: i32) -> Result<(), Box<dyn error::Error>> {
         let value: Value = serde_json::from_slice(key)?;
         let key = unsafe {
             let mut uninit = UninitRow::new(self.vtable);
-            (self.deserialize_fn)(uninit.as_mut_ptr(), &value);
+            call_deserialize_fn(self.deserialize_fn, uninit.as_mut_ptr(), &value)?;
             uninit.assume_init()
         };
 
@@ -253,11 +268,7 @@ impl DbspCircuit {
             )
         };
 
-        Some(JsonSetHandle {
-            handle,
-            deserialize_fn,
-            vtable,
-        })
+        Some(JsonSetHandle::new(handle, deserialize_fn, vtable))
     }
 
     pub fn append_input(&mut self, target: NodeId, data: &StreamCollection) {
@@ -318,7 +329,11 @@ impl DbspCircuit {
 
     // TODO: We probably want other ways to ingest json, e.g. `&[u8]`, `R: Read`,
     // etc.
-    pub fn append_json_input<R>(&mut self, target: NodeId, json: R)
+    pub fn append_json_input<R>(
+        &mut self,
+        target: NodeId,
+        json: R,
+    ) -> Result<(), Box<dyn error::Error>>
     where
         R: Read,
     {
@@ -343,11 +358,10 @@ impl DbspCircuit {
                     let mut batch = Vec::new();
                     let stream = Deserializer::from_reader(json).into_iter::<Value>();
                     for value in stream {
-                        // FIXME: Error handling
-                        let value = value.unwrap();
-
+                        let value = value?;
                         let mut row = UninitRow::new(key_vtable);
-                        unsafe { deserialize_json(row.as_mut_ptr(), &value) }
+                        unsafe { call_deserialize_fn(deserialize_json, row.as_mut_ptr(), &value)? }
+
                         batch.push((unsafe { row.assume_init() }, 1));
                     }
 
@@ -366,9 +380,15 @@ impl DbspCircuit {
         } else {
             tracing::info!("appended json to source {target} which is unused, doing nothing");
         }
+
+        Ok(())
     }
 
-    pub fn append_json_record(&mut self, target: NodeId, record: &[u8]) {
+    pub fn append_json_record(
+        &mut self,
+        target: NodeId,
+        record: &[u8],
+    ) -> Result<(), Box<dyn error::Error>> {
         let (input, layout) = self.inputs.get_mut(&target).unwrap_or_else(|| {
             panic!("attempted to append to {target}, but {target} is not a source node or doesn't exist");
         });
@@ -387,9 +407,10 @@ impl DbspCircuit {
                         )
                     };
 
-                    let value = serde_json::from_slice::<Value>(record).unwrap();
+                    let value = serde_json::from_slice::<Value>(record)?;
                     let mut row = UninitRow::new(key_vtable);
-                    unsafe { deserialize_json(row.as_mut_ptr(), &value) }
+                    unsafe { call_deserialize_fn(deserialize_json, row.as_mut_ptr(), &value)? }
+
                     input
                         .as_set_mut()
                         .unwrap()
@@ -406,6 +427,8 @@ impl DbspCircuit {
         } else {
             tracing::info!("appended json to source {target} which is unused, doing nothing");
         }
+
+        Ok(())
     }
 
     pub fn append_csv_input(&mut self, target: NodeId, path: &Path) {
@@ -600,352 +623,5 @@ unsafe fn constant_from_column(
         )),
 
         ColumnType::Ptr => todo!(),
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::{
-        codegen::CodegenConfig,
-        facade::Demands,
-        ir::{
-            literal::{NullableConstant, RowLiteral, StreamCollection},
-            nodes::{IndexByColumn, StreamKind, StreamLayout},
-            ColumnType, Constant, Graph, GraphExt, NodeId, RowLayoutBuilder,
-        },
-        sql_graph::SqlGraph,
-        utils, DbspCircuit,
-    };
-    use std::path::Path;
-
-    #[test]
-    fn time_series_enrich_e2e() {
-        utils::test_logger();
-
-        // Deserialize the graph from json
-        let graph = serde_json::from_str::<SqlGraph>(TIME_SERIES_ENRICH_SRC)
-            .unwrap()
-            .rematerialize();
-
-        let transactions_layout = graph.nodes()[&TRANSACTIONS_ID]
-            .clone()
-            .unwrap_source()
-            .layout();
-        let demographics_layout = graph.nodes()[&DEMOGRAPHICS_ID]
-            .clone()
-            .unwrap_source()
-            .layout();
-
-        let mut demands = Demands::new();
-        demands.add_csv_deserialize(transactions_layout, transaction_mappings());
-        demands.add_csv_deserialize(demographics_layout, demographic_mappings());
-
-        // Create the circuit
-        let mut circuit = DbspCircuit::new(graph, true, 1, CodegenConfig::debug(), demands);
-
-        // Ingest data
-        circuit.append_csv_input(
-            TRANSACTIONS_ID,
-            &Path::new(PATH).join("transactions_20K.csv"),
-        );
-        circuit.append_csv_input(DEMOGRAPHICS_ID, &Path::new(PATH).join("demographics.csv"));
-
-        // Step the circuit
-        circuit.step().unwrap();
-
-        // TODO: Inspect outputs
-        let _output = circuit.consolidate_output(SINK_ID);
-
-        // Shut down the circuit
-        circuit.kill().unwrap();
-    }
-
-    #[test]
-    fn time_series_enrich_e2e_2() {
-        utils::test_logger();
-
-        // Deserialize the graph from json
-        let mut graph = Graph::new();
-
-        let unit_layout = graph.layout_cache().unit();
-        let demographics_layout = graph.layout_cache().add(
-            RowLayoutBuilder::new()
-                .with_column(ColumnType::F64, false)
-                .with_column(ColumnType::String, true)
-                .with_column(ColumnType::String, true)
-                .with_column(ColumnType::String, true)
-                .with_column(ColumnType::String, true)
-                .with_column(ColumnType::String, true)
-                .with_column(ColumnType::I32, true)
-                .with_column(ColumnType::F64, true)
-                .with_column(ColumnType::F64, true)
-                .with_column(ColumnType::I32, true)
-                .with_column(ColumnType::String, true)
-                .with_column(ColumnType::Date, true)
-                .build(),
-        );
-        let transactions_layout = graph.layout_cache().add(
-            RowLayoutBuilder::new()
-                .with_column(ColumnType::Timestamp, false)
-                .with_column(ColumnType::F64, false)
-                .with_column(ColumnType::String, true)
-                .with_column(ColumnType::String, true)
-                .with_column(ColumnType::F64, true)
-                .with_column(ColumnType::String, true)
-                .with_column(ColumnType::I32, true)
-                .with_column(ColumnType::F64, true)
-                .with_column(ColumnType::F64, true)
-                .with_column(ColumnType::I32, true)
-                .build(),
-        );
-        let key_layout = graph.layout_cache().add(
-            RowLayoutBuilder::new()
-                .with_column(ColumnType::F64, false)
-                .build(),
-        );
-        let culled_demographics_layout = graph.layout_cache().add(
-            RowLayoutBuilder::new()
-                .with_column(ColumnType::String, true)
-                .with_column(ColumnType::String, true)
-                .build(),
-        );
-        let culled_transactions_layout = graph.layout_cache().add(
-            RowLayoutBuilder::new()
-                .with_column(ColumnType::Timestamp, false)
-                .build(),
-        );
-        let output_layout = graph.layout_cache().add(
-            RowLayoutBuilder::new()
-                .with_column(ColumnType::Timestamp, false)
-                .with_column(ColumnType::F64, false)
-                .with_column(ColumnType::String, true)
-                .with_column(ColumnType::String, true)
-                .build(),
-        );
-
-        let demographics_src = graph.source(demographics_layout);
-        let transactions_src = graph.source(transactions_layout);
-
-        let indexed_demographics = graph.add_node(IndexByColumn::new(
-            demographics_src,
-            demographics_layout,
-            0,
-            vec![2, 3, 5, 6, 7, 8, 9, 10, 11],
-            key_layout,
-            culled_demographics_layout,
-        ));
-        let indexed_transactions = graph.add_node(IndexByColumn::new(
-            transactions_src,
-            transactions_layout,
-            1,
-            vec![2, 3, 4, 5, 6, 7, 8, 9],
-            key_layout,
-            culled_transactions_layout,
-        ));
-
-        let transactions_join_demographics = graph.join_core(
-            indexed_transactions,
-            indexed_demographics,
-            {
-                let mut builder = graph.function_builder();
-
-                let key = builder.add_input(key_layout);
-                let transaction = builder.add_input(culled_transactions_layout);
-                let demographic = builder.add_input(culled_demographics_layout);
-                let output = builder.add_output(output_layout);
-                let _unit_output = builder.add_output(unit_layout);
-
-                let trans_date_trans_time = builder.load(transaction, 0);
-                let cc_num = builder.load(key, 0);
-                builder.store(output, 0, trans_date_trans_time);
-                builder.store(output, 1, cc_num);
-
-                {
-                    let first_not_null = builder.create_block();
-                    let after = builder.create_block();
-
-                    let first_null = builder.is_null(demographic, 0);
-                    builder.set_null(output, 2, first_null);
-                    builder.branch(first_null, after, [], first_not_null, []);
-
-                    builder.move_to(first_not_null);
-                    let first = builder.load(demographic, 0);
-                    let first = builder.copy(first);
-                    builder.store(output, 2, first);
-                    builder.jump(after, []);
-
-                    builder.move_to(after);
-                }
-
-                {
-                    let city_not_null = builder.create_block();
-                    let after = builder.create_block();
-
-                    let city_null = builder.is_null(demographic, 1);
-                    builder.set_null(output, 3, city_null);
-                    builder.branch(city_null, after, [], city_not_null, []);
-
-                    builder.move_to(city_not_null);
-                    let city = builder.load(demographic, 1);
-                    let city = builder.copy(city);
-                    builder.store(output, 3, city);
-                    builder.jump(after, []);
-
-                    builder.move_to(after);
-                }
-
-                builder.ret_unit();
-                builder.build()
-            },
-            output_layout,
-            unit_layout,
-            StreamKind::Set,
-        );
-
-        let sink = graph.sink(
-            transactions_join_demographics,
-            StreamLayout::Set(output_layout),
-        );
-
-        let mut demands = Demands::new();
-        demands.add_csv_deserialize(transactions_layout, transaction_mappings());
-        demands.add_csv_deserialize(demographics_layout, demographic_mappings());
-
-        // Create the circuit
-        let mut circuit = DbspCircuit::new(graph, true, 1, CodegenConfig::debug(), demands);
-
-        // Ingest data
-        circuit.append_csv_input(
-            transactions_src,
-            &Path::new(PATH).join("transactions_20K.csv"),
-        );
-        circuit.append_csv_input(demographics_src, &Path::new(PATH).join("demographics.csv"));
-
-        // Step the circuit
-        circuit.step().unwrap();
-
-        // TODO: Inspect outputs
-        let _output = circuit.consolidate_output(sink);
-
-        // Shut down the circuit
-        circuit.kill().unwrap();
-    }
-
-    const PATH: &str = concat!(
-        env!("CARGO_MANIFEST_DIR"),
-        "/../../demo/project_demo01-TimeSeriesEnrich",
-    );
-
-    fn transaction_mappings() -> Vec<(usize, usize, Option<String>)> {
-        vec![
-            (0, 0, Some("%F %T".into())),
-            (1, 1, None),
-            (2, 2, None),
-            (3, 3, None),
-            (4, 4, None),
-            (5, 5, None),
-            (6, 6, None),
-            (7, 7, None),
-            (8, 8, None),
-            (9, 9, None),
-        ]
-    }
-
-    fn demographic_mappings() -> Vec<(usize, usize, Option<String>)> {
-        vec![
-            (0, 0, None),
-            (1, 1, None),
-            (2, 2, None),
-            (3, 3, None),
-            (4, 4, None),
-            (5, 5, None),
-            (6, 6, None),
-            (7, 7, None),
-            (8, 8, None),
-            (9, 9, None),
-            (10, 10, None),
-            (11, 11, Some("%F".into())),
-        ]
-    }
-
-    const TRANSACTIONS_ID: NodeId = NodeId::new(54);
-    const DEMOGRAPHICS_ID: NodeId = NodeId::new(68);
-    const SINK_ID: NodeId = NodeId::new(273);
-
-    static TIME_SERIES_ENRICH_SRC: &str = include_str!("time_series_enrich.json");
-    static CONSTANT_STREAM_TEST: &str = include_str!("constant_stream.json");
-    static UNUSED_SOURCE: &str = include_str!("unused_source.json");
-
-    #[test]
-    fn constant_stream() {
-        utils::test_logger();
-
-        // Deserialize the graph from json
-        let graph = serde_json::from_str::<SqlGraph>(CONSTANT_STREAM_TEST)
-            .unwrap()
-            .rematerialize();
-
-        // Create the circuit
-        let mut circuit = DbspCircuit::new(graph, true, 1, CodegenConfig::debug(), Demands::new());
-
-        // Step the circuit
-        circuit.step().unwrap();
-
-        // Inspect outputs
-        let output = circuit.consolidate_output(NodeId::new(2));
-
-        // Shut down the circuit
-        circuit.kill().unwrap();
-
-        // Ensure the output is correct
-        let expected = StreamCollection::Set(vec![(
-            RowLiteral::new(vec![NullableConstant::NonNull(Constant::U32(1))]),
-            1,
-        )]);
-        assert_eq!(output, expected);
-    }
-
-    #[test]
-    fn append_unused_source() {
-        utils::test_logger();
-
-        // Deserialize the graph from json
-        let graph = serde_json::from_str::<SqlGraph>(UNUSED_SOURCE)
-            .unwrap()
-            .rematerialize();
-
-        // Create the circuit
-        let mut circuit = DbspCircuit::new(graph, true, 1, CodegenConfig::debug(), Demands::new());
-
-        // Feed data to our unused input
-        circuit.append_input(
-            NodeId::new(1),
-            &StreamCollection::Set(vec![(
-                RowLiteral::new(vec![
-                    NullableConstant::NonNull(Constant::I32(1)),
-                    NullableConstant::NonNull(Constant::F64(1.0)),
-                    NullableConstant::NonNull(Constant::Bool(true)),
-                    NullableConstant::NonNull(Constant::String("foobar".into())),
-                    NullableConstant::Nullable(Some(Constant::I32(1))),
-                    NullableConstant::Nullable(Some(Constant::F64(1.0))),
-                ]),
-                1,
-            )]),
-        );
-
-        // Step the circuit
-        circuit.step().unwrap();
-
-        let output = circuit.consolidate_output(NodeId::new(3));
-
-        // Kill the circuit
-        circuit.kill().unwrap();
-
-        // Ensure the output is correct
-        let expected = StreamCollection::Set(vec![(
-            RowLiteral::new(vec![NullableConstant::NonNull(Constant::I32(0))]),
-            1,
-        )]);
-        assert_eq!(output, expected);
     }
 }
